# src/docs/swagger.yml
openapi: 2.0
info:
  title: API - Jhonathan
  description: "API para gerenciar usuários, produtos, categorias e pedidos."
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Lista
      operationId: getProduct
      responses:
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "400":
          description: Requisição inválida
        "404":
          description: Não encontrado
        "500":
          description: Servidor não disponível
    post:
      tags:
        - Products
      summary: Cria um novo produto
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  "/products/{id}":
    get:
      tags:
        - Users
      summary: Busca um produto por ID
      operationId: getUserById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Produto não encontrado
    put:
      tags:
        - Products
      summary: Atualiza um produto por ID
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Produto atualizado
    delete:
      tags:
        - Products
      summary: Deleta um produto por ID
      operationId: deleteProduct
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Produto deletado com sucesso
...
  "/categorys/{id}":
    get:
      tags:
        - Categorys
      summary: Busca uma categoria por ID
      operationId: getUserById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: categoria encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: categoria não encontrada
    put:
      tags:
        - Categorys
      summary: Atualiza uma categoria por ID
      operationId: updateCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        "200":
          description: categoria atualizada
    delete:
      tags:
        - Categorys
      summary: Deleta uma categoria por ID
      operationId: deleteCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: categoria deletada com sucesso
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Digite o token no formato: Bearer <token>"
security:
  - BearerAuth: []